djangoStudentkit/pyproject.toml
[tool.poetry]
name = "djangostudentkit"
version = "0.1.0"
description = ""
authors = ["eduetopa <eduetopa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.14.0"
mysqlclient = "^2.1.1"
mysql = "^0.0.3"
django-cors-headers = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
